account-service/requirements.txt
django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
psycopg2-binary==2.9.*
python-decouple==3.8
kafka-python==2.0.*
minio==7.2.*
drf-spectacular==0.27.*
pytest==8.2.*
pytest-django==4.8.*
pytest-cov==5.*
factory-boy==3.3.*
black==24.*
isort==5.*
flake8==7.*
mypy==1.*
django-stubs==5.*